+++ examples/nestjs-basic/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prismakit/example-nestjs-basic
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @prismakit/core@2.1.1
+  - @prismakit/nestjs@2.1.1
+
 ## 0.0.2
 
 ### Patch Changes