+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prismakit/eslint-plugin
 
+## 2.1.1
+
+### Patch Changes
+
+- AutoComposer injects the target primary key into nested relation selects so compose mapping works even when callers omit `id`.
+
 ## 2.1.0
 
 ### Minor Changes